service/backend/service/fastapi-server/src/llm/requirements.txt
# FastAPI & Server
fastapi>=0.95.0
uvicorn>=0.21.1

# DB & ORM
sqlalchemy>=2.0.7
psycopg2-binary>=2.9.5

# Validation & Security
pydantic>=1.10.7
email-validator>=2.0.0
passlib>=1.7.4
bcrypt>=4.0.1
python-jose[cryptography]>=3.3.0

# Form & Auth
python-multipart>=0.0.6

# HTTP & Env
httpx>=0.24.0
python-dotenv>=1.0.0

# AI / LLM
openai>=1.0.0
langchain==0.1.17
langchain-core==0.1.52
langchain-openai==0.0.8
langchain-community==0.0.38
langsmith>=0.0.60
faiss-cpu>=1.7.4

# Data
pandas>=2.0.0